+++ src/upgradecontrol.cpp
@@ -26,7 +26,7 @@
 int32_t UpgradeControl_Factors[UPGRADE_CONTROL_COUNT] = {4, 4, 8, 2, 2, 2, 8, 4, 16};
 
 UpgradeControl::UpgradeControl(WinID window_id, int32_t ulx, int32_t uly, int32_t button_right_r_value,
-                               int32_t button_left_r_value, uint16_t *team_gold)
+                               int32_t button_left_r_value, int32_t *team_gold)
     : team_gold(team_gold), factor(1), upgrade_ratio(nullptr), uly(uly) {
     upgrade_right = new (std::nothrow) Button(UPGRGT_U, UPGRGT_D, ulx + 18, uly);
     upgrade_right->SetRValue(button_right_r_value);
@@ -91,7 +91,7 @@ int32_t UpgradeControl::GetRatio(uint16_t value) const {
     return result;
 }
 
-void UpgradeControl::SetupRatio(uint16_t factor, uint16_t *upgrade_ratio) {
+void UpgradeControl::SetupRatio(uint16_t factor, int32_t *upgrade_ratio) {
     this->factor = factor;
     this->upgrade_ratio = upgrade_ratio;
 }
